feat : create index page

+++ attendance/src/main/java/smmiddle/attendance/repository/AttendanceRepository.java
@@ -3,7 +3,7 @@ package smmiddle.attendance.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import smmiddle.attendance.entity.Attendance;
 
-public interface AttendanceRepository extends JpaRepository<Attendance, Integer> {
+public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
 
 
 }
