+++ attendance/src/main/java/smmiddle/attendance/repository/CellRepository.java
@@ -3,6 +3,6 @@ package smmiddle.attendance.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import smmiddle.attendance.entity.Cell;
 
-public interface CellRepository extends JpaRepository<Cell, Integer> {
+public interface CellRepository extends JpaRepository<Cell, Long> {
 
 }
